build: image size compression client side

+++ controllers/deleteController.js
@@ -8,15 +8,16 @@ export async function deleteGet(req, res) {
   //------delete uploaded img--------
   if (imgfile && imgfile.length > 0) {
     const imgPath = process.cwd() + `/public/uploads/${imgfile}`;
-    await fs.rm(imgPath, { recursive: true }, (err) => {
-      if (err) {
-        console.error(err.message);
-        return;
-      }
-      console.log("File deleted successfully");
-    });
+    if (imgPath) {
+      await fs.rm(imgPath, { recursive: true }, (err) => {
+        if (err) {
+          console.error(err.message);
+          return;
+        }
+        console.log("File deleted successfully");
+      });
+    }
   }
-
   await deletePost(id);
   res.redirect("/");
 }
